WIP: New tests verifying that data is transfered after filter
Addition of tests that make sure that when converting Dream3D data
to ITK data and ITK data to Dream3D data, the resulting data
container belongs to the new object, and is not destroyed when the
original object is destroyed. Currently, these tests are failing
because the ownership of the data is not transfered.