BUG: Reverting static_cast to dynamic_cast
In the Graft() function in itk::Dream3DImage, dynamic_cast had been
replaced by itkDynamicCastInDebugMode (static_cast) to  prevent some
errors on some compilers which do not support the use of dynamic_cast
with templated classes instantiated in a translation unit different than
the one where they are defined [1] in commit
74a125e4787aad826a9ca83f7cc32f0d527c65ac . However, if a developer
provides a DataObject to the Graft() function that is not an
itkDream3DImage, this will lead to a crash. To avoid this problem,
we revert the code to dynamic_cast.

[1] https://github.com/Slicer/Slicer/pull/573#issuecomment-244484002