ENH: Adding a virtual class to develop ITK based filters.
Dream3D ITK filter classes are based on ITKImageBase class
which implements:
* dataCheck() that verifies the general inputs (not specific
to a filter) and create a new array in the given container
if that option is selected.
* filter() that converts a data container to an ITK image,
runs and ITK filter (expects ITKImageToImageFilter) and
converts the result back to a data container.
* execute() and preflight() are also implemented and call
the virtual functions dataCheckInternal() and filterInternal().
These two functions are pure virtual functions in ITKImageBase
and need to be overloaded in the actual filter implementation.
The overriding methods are expected to use Dream3DArraySwitchMacro
to call the corresponding templated method implemented in
ITKImageBase.