BUG: Missing ITK IO added
ITKIOBMP and ITKIOJPEG were not required as SCIFIO could read
and write these types of images. However, SCIFIO relies on
the java virtual machine which slows down the process. To
speed up reading and writing these two types of images, ITKIOBMP
and ITKIOJPEG are used instead of SCIFIO.
ITKIOIPL was not required in the CMakeLists.txt. This lead to
the plugin not being loaded (missing dependency) on Windows.